Implement api key access

+++ src/main/java/fr/cesi/meteo/application/presenter/api/IndexController.java
@@ -1,6 +1,7 @@
 package fr.cesi.meteo.application.presenter.api;
 
 import fr.cesi.meteo.configuration.factory.ServiceFactory;
+import fr.cesi.meteo.domain.service.IApiKeyService;
 import fr.cesi.meteo.domain.service.IDataService;
 import fr.cesi.meteo.domain.service.IResponseService;
 import fr.cesi.meteo.infrastructure.http.Method;
@@ -10,10 +11,16 @@ import fr.cesi.meteo.infrastructure.http.annotation.Action;
 
 public class IndexController {
 
+    private final IDataService dataService;
+    private final IApiKeyService apiKeyService;
+
+    public IndexController() {
+        this.dataService = ServiceFactory.getInstance().getDataService();
+        this.apiKeyService = ServiceFactory.getInstance().getApiKeyService();
+    }
+
     @Action(path = "/", method = Method.GET)
     public Response getIndexAction(Request request, Response response) {
-        IDataService dataService = ServiceFactory.getInstance().getDataService();
-
         response.setStatusCode(200);
         response.setBody(dataService.getDataCollection(request));
 
@@ -22,7 +29,11 @@ public class IndexController {
 
     @Action(path = "/create", method = Method.POST)
     public Response postIndexAction(Request request, Response response) {
-        IDataService dataService = ServiceFactory.getInstance().getDataService();
+        if (!apiKeyService.verifyApiToken(request)) {
+            response.setStatusCode(403);
+            return response;
+        }
+
         IResponseService responseService = ServiceFactory.getInstance().getResponseService();
 
         if (dataService.addNewData(request)) {
@@ -38,7 +49,11 @@ public class IndexController {
 
     @Action(path = "/update", method = Method.PUT)
     public Response updateIndexAction(Request request, Response response) {
-        IDataService dataService = ServiceFactory.getInstance().getDataService();
+        if (!apiKeyService.verifyApiToken(request)) {
+            response.setStatusCode(403);
+            return response;
+        }
+
         IResponseService responseService = ServiceFactory.getInstance().getResponseService();
 
         if (dataService.updateData(request))
@@ -51,7 +66,10 @@ public class IndexController {
 
     @Action(path = "/delete", method = Method.DELETE)
     public Response deleteIndexAction(Request request, Response response) {
-        IDataService dataService = ServiceFactory.getInstance().getDataService();
+        if (!apiKeyService.verifyApiToken(request)) {
+            response.setStatusCode(403);
+            return response;
+        }
 
         if (dataService.deleteData(request))
             response.setStatusCode(200);
